Model sizes should be ushorts

+++ Voxel2Pixel/Interfaces/IModel.cs
@@ -2,9 +2,9 @@
 {
 	public interface IModel : IFetch
 	{
-		int SizeX { get; }
-		int SizeY { get; }
-		int SizeZ { get; }
+		ushort SizeX { get; }
+		ushort SizeY { get; }
+		ushort SizeZ { get; }
 		bool IsInside(int x, int y, int z);
 		bool IsOutside(int x, int y, int z);
 	}
