+++ Voxel2Pixel/Model/ArrayModel.cs
@@ -2,7 +2,7 @@
 
 namespace Voxel2Pixel.Model
 {
-    public class ArrayModel : IModel, ITurnable
+	public class ArrayModel : IModel, ITurnable
 	{
 		public ArrayModel(byte[][][] voxels) => Voxels = voxels;
 		public ArrayModel(ArrayModel other) : this(other.Voxels.DeepCopy()) { }
@@ -17,9 +17,9 @@ namespace Voxel2Pixel.Model
 		}
 		public byte[][][] Voxels { get; set; }
 		#region IModel
-		public int SizeX => Voxels.Length;
-		public int SizeY => Voxels[0].Length;
-		public int SizeZ => Voxels[0][0].Length;
+		public ushort SizeX => (ushort)Voxels.Length;
+		public ushort SizeY => (ushort)Voxels[0].Length;
+		public ushort SizeZ => (ushort)Voxels[0][0].Length;
 		public byte? At(int x, int y, int z) => IsInside(x, y, z) ? Voxels[x][y][z] : (byte?)null;
 		public bool IsInside(int x, int y, int z) => !IsOutside(x, y, z);
 		public bool IsOutside(int x, int y, int z) => x < 0 || y < 0 || z < 0 || x >= SizeX || y >= SizeY || z >= SizeZ;
