+++ Voxel2Pixel/Model/TurnModel.cs
@@ -3,16 +3,16 @@ using Voxel2Pixel.Interfaces;
 
 namespace Voxel2Pixel.Model
 {
-    /// <summary>
-    /// Rotates models at 90 degree angles, including their sizes
-    /// </summary>
-    public class TurnModel : ContainerModel, ITurnable
+	/// <summary>
+	/// Rotates models at 90 degree angles, including their sizes
+	/// </summary>
+	public class TurnModel : ContainerModel, ITurnable
 	{
 		public CuboidOrientation CuboidOrientation { get; set; } = CuboidOrientation.SOUTH0;
 		#region ContainerModel
-		public override int SizeX => RotatedSize(0);
-		public override int SizeY => RotatedSize(1);
-		public override int SizeZ => RotatedSize(2);
+		public override ushort SizeX => RotatedSize(0);
+		public override ushort SizeY => RotatedSize(1);
+		public override ushort SizeZ => RotatedSize(2);
 		public override bool IsInside(int x, int y, int z) => !IsOutside(x, y, z);
 		public override bool IsOutside(int x, int y, int z) => x < 0 || y < 0 || z < 0 || x >= SizeX || y >= SizeY || z >= SizeZ;
 		public override byte? At(int x, int y, int z)
@@ -82,13 +82,13 @@ namespace Voxel2Pixel.Model
 		}
 		#endregion ReverseRotate
 		#region Size
-		public int RotatedSize(int axis) => ModelSize(CuboidOrientation.ReverseAffected(axis));
+		public ushort RotatedSize(int axis) => ModelSize(CuboidOrientation.ReverseAffected(axis));
 		/// <summary>
 		/// Allows treating the sizes of the underlying model as if they were in an array. Allows negative values for index, unlike an array, and will correctly treat the negative index that corresponds to a reversed axis as if it was the the corresponding non-reversed axis. This means -1 will be the same as 0, -2 the same as 1, and -3 the same as 2.
 		/// </summary>
 		/// <param name="index">index 0 or -1 for x, 1 or -2 for y, or 2 or -3 for z; negative index values have the same size, but different starts and directions</param>
 		/// <returns>the size of the specified dimension</returns>
-		public int ModelSize(int index)
+		public ushort ModelSize(int index)
 		{
 			switch (index)
 			{
